refactor(buffer/join): Improve memory allocation efficiency in split_batch
